agrobot_api/requirements.txt
alembic==1.0.6
aniso8601==4.1.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
protobuf==3.6.1
PyMySQL==0.10.0
psycopg2==2.8.5
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.18
uWSGI==2.0.18
Werkzeug==0.14.1
paho-mqtt==1.4.0
sqlalchemy-utils==0.36.8
